Add database operations for plans, postprocess jobs, projects, and settings
- Implemented PlanOperationsMixin for managing plans and tasks, including creation, retrieval, updating, and deletion.
- Added PostprocessOperationsMixin to handle postprocess job lifecycle, including job creation, status updates, and metrics tracking.
- Introduced ProjectOperationsMixin for project management, including project creation, retrieval, and linking chats to projects.
- Created SettingsOperationsMixin for managing user preferences, volume metadata, memory configuration, and MCP server operations.
- Each mixin includes methods for interacting with the database using SQLModel and handling various operations related to their respective domains.
