--- src/suzent/database.py
@@ -1,3031 +1,0 @@
-"""
-Database layer for chat persistence using SQLModel.
-
-Provides a clean, type-safe interface for all database operations.
-"""
-
-import base64
-import json
-import os
-import uuid
-from datetime import datetime, timedelta, timezone
-from pathlib import Path
-from typing import Any, Dict, List, Optional
-from pydantic import BaseModel
-from sqlalchemy.orm import selectinload
-from sqlalchemy import cast, Text, or_, text, inspect, func
-from sqlalchemy.orm.attributes import flag_modified
-from sqlmodel import (
-    Column,
-    Field,
-    JSON,
-    Relationship,
-    Session,
-    SQLModel,
-    create_engine,
-    select,
-)
-
-
-def _json_escape_for_like(s: str) -> str:
-    """Return the \\uXXXX form of any non-ASCII chars in s.
-
-    SQLAlchemy's JSON column stores non-ASCII characters as \\uXXXX escape
-    sequences (ensure_ascii=True default). A plain LIKE '%中文%' therefore
-    never matches; we must search for '\\u4e2d\\u6587' instead.
-    ASCII characters are left unchanged so English searches still work.
-    """
-    result = []
-    for ch in s:
-        if ord(ch) > 127:
-            result.append(f"\\u{ord(ch):04x}")
-        else:
-            result.append(ch)
-    return "".join(result)
-
-
-def _messages_search_filter(search: str):
-    """Build an OR filter that matches search in the messages JSON column.
-
-    Searches both the raw text (for any future ensure_ascii=False storage)
-    and the \\uXXXX-escaped form (for current ensure_ascii=True storage).
-    """
-    escaped = _json_escape_for_like(search)
-    if escaped == search:
-        return cast(ChatModel.messages, Text).contains(search)
-    return or_(
-        cast(ChatModel.messages, Text).contains(search),
-        cast(ChatModel.messages, Text).contains(escaped),
-    )
-
-
-# -----------------------------------------------------------------------------
-# SQLModel Table Definitions
-# -----------------------------------------------------------------------------
-
-
-class ChatSummaryModel(BaseModel):
-    """Bail-out model for chat listing."""
-
-    id: str
-    title: str
-    createdAt: str
-    updatedAt: str
-    messageCount: int
-    lastMessage: Optional[str] = None
-    platform: Optional[str] = None
-    heartbeatEnabled: bool = False
-    lastResultAt: Optional[str] = None
-    unreadCount: int = 0
-    projectId: Optional[str] = None
-    projectSlug: Optional[str] = None
-    projectName: Optional[str] = None
-    parentChatId: Optional[str] = None
-
-
-class ProjectModel(SQLModel, table=True):
-    """Project groups multiple chat sessions and owns a shared plan."""
-
-    __tablename__ = "projects"
-
-    id: str = Field(primary_key=True, default_factory=lambda: str(uuid.uuid4()))
-    name: str
-    slug: str = Field(unique=True, index=True)
-    created_at: datetime = Field(default_factory=lambda: datetime.now(timezone.utc))
-    archived: bool = Field(default=False)
-
-    chats: List["ChatModel"] = Relationship(back_populates="project")
-
-
-class ChatModel(SQLModel, table=True):
-    """Chat session with messages and configuration."""
-
-    __tablename__ = "chats"
-
-    id: str = Field(primary_key=True)
-    title: str
-    created_at: datetime = Field(serialization_alias="createdAt")
-    updated_at: datetime = Field(serialization_alias="updatedAt")
-    config: dict = Field(default_factory=dict, sa_column=Column(JSON))
-    messages: list = Field(default_factory=list, sa_column=Column(JSON))
-    context_usage: dict = Field(default_factory=dict, sa_column=Column(JSON))
-    agent_state: Optional[bytes] = None
-
-    # Session lifecycle fields
-    last_active_at: Optional[datetime] = None
-    turn_count: int = Field(default=0)
-
-    # Two-phase post-process state tracking
-    state_revision: int = Field(default=0)
-    finalized_revision: int = Field(default=0)
-    state_stage: Optional[str] = None  # snapshot | finalized
-    state_updated_at: Optional[datetime] = None
-
-    # Background execution tracking — written only by background executors
-    last_result_at: Optional[datetime] = None
-
-    # Working directory binding (S2O Phase 1)
-    working_directory: Optional[str] = None
-
-    # Project association
-    project_id: Optional[str] = Field(
-        default=None, foreign_key="projects.id", index=True
-    )
-    project: Optional[ProjectModel] = Relationship(back_populates="chats")
-
-    plans: List["PlanModel"] = Relationship(
-        back_populates="chat",
-        sa_relationship_kwargs={"cascade": "all, delete-orphan"},
-    )
-
-
-class RetryCheckpointModel(SQLModel, table=True):
-    """Stores the last retry checkpoint for a chat session (one row per chat)."""
-
-    __tablename__ = "retry_checkpoints"
-
-    chat_id: str = Field(primary_key=True)
-    agent_state_before: Optional[bytes] = None
-    messages_before: list = Field(default_factory=list, sa_column=Column(JSON))
-    user_message: str = Field(default="")
-    user_files: list = Field(default_factory=list, sa_column=Column(JSON))
-    config_snapshot: dict = Field(default_factory=dict, sa_column=Column(JSON))
-    has_file_snapshot: bool = Field(default=False)
-    # Serialised FileTracker snapshot: list of {path, backup_name, version, backup_time}
-    file_snapshot: list = Field(default_factory=list, sa_column=Column(JSON))
-    created_at: datetime = Field(default_factory=lambda: datetime.now(timezone.utc))
-
-
-class PlanModel(SQLModel, table=True):
-    """Execution plan associated with a chat session."""
-
-    __tablename__ = "plans"
-
-    id: Optional[int] = Field(default=None, primary_key=True)
-    chat_id: str = Field(foreign_key="chats.id", index=True)
-    objective: str
-    created_at: datetime = Field(serialization_alias="createdAt")
-    updated_at: datetime = Field(serialization_alias="updatedAt")
-
-    chat: Optional[ChatModel] = Relationship(back_populates="plans")
-    tasks: List["TaskModel"] = Relationship(
-        back_populates="plan",
-        sa_relationship_kwargs={"cascade": "all, delete-orphan"},
-    )
-
-
-class TaskModel(SQLModel, table=True):
-    """Individual task within a plan."""
-
-    __tablename__ = "tasks"
-
-    id: Optional[int] = Field(default=None, primary_key=True)
-    plan_id: int = Field(foreign_key="plans.id", index=True)
-    number: int = Field(index=True)
-    description: str
-    status: str = Field(default="pending")
-    note: Optional[str] = None
-    capabilities: Optional[str] = None
-    created_at: datetime = Field(serialization_alias="createdAt")
-    updated_at: datetime = Field(serialization_alias="updatedAt")
-
-    plan: Optional[PlanModel] = Relationship(back_populates="tasks")
-
-
-class UserPreferencesModel(SQLModel):
-    """Singleton table for global user preferences."""
-
-    id: int = Field(default=1, primary_key=True)
-    model: Optional[str] = None
-    agent: Optional[str] = None
-    tools: Optional[list] = Field(default=None, sa_column=Column(JSON))
-    memory_enabled: bool = Field(default=False)
-    sandbox_enabled: bool = Field(default=True)
-    sandbox_volumes: Optional[list] = Field(default=None, sa_column=Column(JSON))
-    updated_at: datetime = Field(serialization_alias="updatedAt")
-
-
-class VolumeMetadataModel(SQLModel, table=True):
-    """Cached metadata for configured custom volumes."""
-
-    __tablename__ = "volume_metadata"
-
-    volume: str = Field(primary_key=True)
-    host_path: str
-    mount_point: str
-    kind: str = Field(default="generic")
-    exists: bool = Field(default=False)
-    is_git_repo: Optional[bool] = None
-    git_root: Optional[str] = None
-    status: str = Field(default="unknown")
-    error: Optional[str] = None
-    checked_at: datetime = Field(serialization_alias="checkedAt")
-
-
-class MCPServerModel(SQLModel, table=True):
-    """MCP server configuration."""
-
-    __tablename__ = "mcp_servers"
-
-    name: str = Field(primary_key=True)
-    type: str  # "url" or "stdio"
-    url: Optional[str] = None
-    headers: Optional[dict] = Field(default=None, sa_column=Column(JSON))
-    command: Optional[str] = None
-    args: Optional[list] = Field(default=None, sa_column=Column(JSON))
-    env: Optional[dict] = Field(default=None, sa_column=Column(JSON))
-    enabled: bool = Field(default=True)
-    created_at: datetime = Field(serialization_alias="createdAt")
-    updated_at: datetime = Field(serialization_alias="updatedAt")
-
-
-class ApiKeyModel(SQLModel):
-    """Compatibility model for API key-shaped config values."""
-
-    key: str = Field(primary_key=True)
-    value: str
-    updated_at: datetime = Field(serialization_alias="updatedAt")
-
-
-class CostLedgerModel(SQLModel, table=True):
-    """Global cost ledger — every LLM call is recorded here.
-
-    Entries survive chat deletion so lifetime spend is always accurate.
-    """
-
-    __tablename__ = "cost_ledger"
-
-    id: Optional[int] = Field(default=None, primary_key=True)
-    chat_id: Optional[str] = None
-    model: str
-    role: str = "primary"
-    input_tokens: int = 0
-    output_tokens: int = 0
-    cache_write_tokens: int = 0
-    cache_read_tokens: int = 0
-    cost_usd: float = 0.0
-    created_at: datetime = Field(default_factory=lambda: datetime.now(timezone.utc))
-
-
-class ChatCostSummaryModel(SQLModel, table=True):
-    """Per-chat cost summary — denormalized for fast per-chat queries.
-
-    Can be rebuilt from ``cost_ledger`` if needed.
-    """
-
-    __tablename__ = "chat_cost_summary"
-
-    chat_id: str = Field(primary_key=True)
-    total_cost_usd: float = 0.0
-    total_input_tokens: int = 0
-    total_output_tokens: int = 0
-    total_cache_write_tokens: int = 0
-    total_cache_read_tokens: int = 0
-    last_updated_at: datetime = Field(
-        default_factory=lambda: datetime.now(timezone.utc)
-    )
-
-
-class MemoryConfigModel(SQLModel):
-    """Singleton table for memory system configuration."""
-
-    id: int = Field(default=1, primary_key=True)
-    embedding_model: Optional[str] = None
-    extraction_model: Optional[str] = None
-    updated_at: datetime = Field(serialization_alias="updatedAt")
-
-
-class CronJobModel(SQLModel, table=True):
-    """Scheduled cron job for automated task execution."""
-
-    __tablename__ = "cron_jobs"
-
-    id: Optional[int] = Field(default=None, primary_key=True)
-    name: str
-    cron_expr: str
-    prompt: str
-    active: bool = Field(default=True)
-    delivery_mode: str = Field(default="announce")  # "announce" | "none"
-    model_override: Optional[str] = None
-    retry_count: int = Field(default=0)
-    last_run_at: Optional[datetime] = None
-    next_run_at: Optional[datetime] = None
-    last_result: Optional[str] = None
-    last_error: Optional[str] = None
-    created_at: datetime = Field(serialization_alias="createdAt")
-    updated_at: datetime = Field(serialization_alias="updatedAt")
-
-
-class CronRunModel(SQLModel, table=True):
-    """History record for a single cron job execution."""
-
-    __tablename__ = "cron_runs"
-
-    id: Optional[int] = Field(default=None, primary_key=True)
-    job_id: int = Field(index=True)
-    started_at: datetime
-    finished_at: Optional[datetime] = None
-    status: str = Field(default="running")  # "running" | "success" | "error"
-    result: Optional[str] = None
-    error: Optional[str] = None
-
-
-class PostprocessJobModel(SQLModel, table=True):
-    """Postprocess job tracking for chat turn completion.
-
-    Tracks the full lifecycle of a postprocess job from snapshot through finalize,
-    including per-step status (B1..B7) and retry attempts.
-    """
-
-    __tablename__ = "postprocess_jobs"
-
-    id: Optional[int] = Field(default=None, primary_key=True)
-    job_id: str = Field(index=True, unique=True)  # UUID-based unique ID
-    chat_id: str = Field(foreign_key="chats.id", index=True)
-    assigned_revision: int  # Revision from A phase snapshot
-    current_revision: Optional[int] = None  # Latest revision (for guard check)
-    status: str = Field(
-        default="pending"
-    )  # pending | running | success | failed | skipped_stale
-    outcome: Optional[str] = None  # success | failed | skipped_stale
-    attempt: int = Field(default=1)
-    max_attempts: int = Field(default=3)
-
-    # Step status tracking (B1..B7)
-    step_status_json: Optional[str] = (
-        None  # JSON dict of {step: {status, error, duration_ms}}
-    )
-
-    # Timing
-    started_at: Optional[datetime] = None
-    finished_at: Optional[datetime] = None
-    duration_ms: Optional[int] = None
-
-    # Error tracking
-    error_class: Optional[str] = None
-    error_message: Optional[str] = None
-
-    # Metadata
-    created_at: datetime = Field(default_factory=lambda: datetime.now())
-    updated_at: datetime = Field(default_factory=lambda: datetime.now())
-
-
-# PostProcess Step Constants
-class PostProcessStep:
-    """Constant definitions for postprocess steps B1..B7."""
-
-    TRANSCRIPT = "B1_transcript"
-    MEMORY = "B2_memory"
-    COMPRESS = "B3_compress"
-    DISPLAY = "B4_display"
-    PERSIST = "B5_persist"
-    LIFECYCLE = "B6_lifecycle"
-    MIRROR = "B7_mirror"
-
-    ALL = [TRANSCRIPT, MEMORY, COMPRESS, DISPLAY, PERSIST, LIFECYCLE, MIRROR]
-
-
-# Post Process Status Constants
-class PostProcessStatus:
-    """Constant definitions for postprocess job statuses."""
-
-    PENDING = "pending"
-    RUNNING = "running"
-    SUCCESS = "success"
-    FAILED = "failed"
-    SKIPPED_STALE = "skipped_stale"
-
-
-class PostProcessOutcome:
-    """Constant definitions for postprocess job outcomes."""
-
-    SUCCESS = "success"
-    FAILED = "failed"
-    SKIPPED_STALE = "skipped_stale"
-
-
-# Step Status Constants per Step
-class StepStatus:
-    """Constant definitions for individual step status."""
-
-    PENDING = "pending"
-    RUNNING = "running"
-    SUCCESS = "success"
-    FAILED = "failed"
-
-
-# Metrics counters (in-memory for now, can be persisted later)
-class PostProcessMetrics:
-    """Container for postprocess metrics."""
-
-    def __init__(self):
-        self.snapshot_committed = 0
-        self.snapshot_failed = 0
-        self.job_started = 0
-        self.job_success = 0
-        self.job_failed = 0
-        self.job_skipped_stale = 0
-        self.total_duration_ms = 0
-
-
-# Global metrics instance
-_postprocess_metrics = PostProcessMetrics()
-
-
-# -----------------------------------------------------------------------------
-# Database Management
-# -----------------------------------------------------------------------------
-
-
-class ChatDatabase:
-    """Handles database operations for chat persistence using SQLModel."""
-
-    def __init__(self, db_path: str = None):
-        if db_path is None:
-            # Use data directory from config if available, otherwise relative to project
-            try:
-                from suzent.config import DATA_DIR
-
-                self.db_path = DATA_DIR / "chats.db"
-            except ImportError:
-                self.db_path = Path(".suzent/chats.db")
-        else:
-            self.db_path = Path(db_path)
-
-        # Ensure parent directory exists
-        self.db_path.parent.mkdir(parents=True, exist_ok=True)
-
-        # If db_path is a directory (Docker mount issue), remove it and create file
-        if self.db_path.is_dir():
-            import shutil
-
-            shutil.rmtree(self.db_path)
-
-        # Create engine with SQLite
-        self.engine = create_engine(
-            f"sqlite:///{self.db_path}",
-            echo=False,
-            connect_args={"check_same_thread": False},
-        )
-
-        # Create all tables
-        SQLModel.metadata.create_all(self.engine)
-
-        # Run migrations for new columns
-        self._run_migrations()
-        self._migrate_static_config_from_db()
-        self._ensure_default_project()
-        self._migrate_legacy_session_dirs()
-
-    def _run_migrations(self):
-        """Run database migrations for new columns."""
-        inspector = inspect(self.engine)
-
-        # Migration: Add 'headers' column to 'mcp_servers' table
-        if "mcp_servers" in inspector.get_table_names():
-            columns = [col["name"] for col in inspector.get_columns("mcp_servers")]
-            if "headers" not in columns:
-                with self.engine.connect() as conn:
-                    conn.execute(
-                        text("ALTER TABLE mcp_servers ADD COLUMN headers TEXT")
-                    )
-                    conn.commit()
-
-        # Migration: Add retry_count and drop legacy is_heartbeat from cron_jobs
-        if "cron_jobs" in inspector.get_table_names():
-            columns = [col["name"] for col in inspector.get_columns("cron_jobs")]
-            with self.engine.connect() as conn:
-                if "retry_count" not in columns:
-                    conn.execute(
-                        text(
-                            "ALTER TABLE cron_jobs ADD COLUMN retry_count INTEGER DEFAULT 0"
-                        )
-                    )
-                if "is_heartbeat" in columns:
-                    conn.execute(text("ALTER TABLE cron_jobs DROP COLUMN is_heartbeat"))
-                conn.commit()
-
-        # Migration: Add session lifecycle columns to 'chats' table
-        if "chats" in inspector.get_table_names():
-            columns = [col["name"] for col in inspector.get_columns("chats")]
-            with self.engine.connect() as conn:
-                if "last_active_at" not in columns:
-                    conn.execute(
-                        text("ALTER TABLE chats ADD COLUMN last_active_at DATETIME")
-                    )
-                if "turn_count" not in columns:
-                    conn.execute(
-                        text(
-                            "ALTER TABLE chats ADD COLUMN turn_count INTEGER DEFAULT 0"
-                        )
-                    )
-                if "last_result_at" not in columns:
-                    conn.execute(
-                        text("ALTER TABLE chats ADD COLUMN last_result_at DATETIME")
-                    )
-                if "working_directory" not in columns:
-                    conn.execute(
-                        text("ALTER TABLE chats ADD COLUMN working_directory TEXT")
-                    )
-                if "state_revision" not in columns:
-                    conn.execute(
-                        text(
-                            "ALTER TABLE chats ADD COLUMN state_revision INTEGER DEFAULT 0"
-                        )
-                    )
-                if "finalized_revision" not in columns:
-                    conn.execute(
-                        text(
-                            "ALTER TABLE chats ADD COLUMN finalized_revision INTEGER DEFAULT 0"
-                        )
-                    )
-                if "state_stage" not in columns:
-                    conn.execute(text("ALTER TABLE chats ADD COLUMN state_stage TEXT"))
-                if "state_updated_at" not in columns:
-                    conn.execute(
-                        text("ALTER TABLE chats ADD COLUMN state_updated_at DATETIME")
-                    )
-                conn.commit()
-
-        # Migration: Add file_snapshot columns to retry_checkpoints
-        if "retry_checkpoints" in inspector.get_table_names():
-            columns = [
-                col["name"] for col in inspector.get_columns("retry_checkpoints")
-            ]
-            with self.engine.connect() as conn:
-                if "has_file_snapshot" not in columns:
-                    conn.execute(
-                        text(
-                            "ALTER TABLE retry_checkpoints ADD COLUMN has_file_snapshot BOOLEAN DEFAULT 0"
-                        )
-                    )
-                if "file_snapshot" not in columns:
-                    conn.execute(
-                        text(
-                            "ALTER TABLE retry_checkpoints ADD COLUMN file_snapshot TEXT"
-                        )
-                    )
-                conn.commit()
-
-        # Migration: Ensure postprocess_jobs table exists (auto-created by SQLModel)
-        # Verify it exists after create_all() runs
-        if "postprocess_jobs" not in inspector.get_table_names():
-            # This should not happen as SQLModel.metadata.create_all() already ran
-            # but adding for safety
-            with self.engine.connect() as conn:
-                conn.execute(
-                    text(
-                        """
-                        CREATE TABLE IF NOT EXISTS postprocess_jobs (
-                            id INTEGER PRIMARY KEY AUTOINCREMENT,
-                            job_id TEXT NOT NULL UNIQUE,
-                            chat_id TEXT NOT NULL,
-                            assigned_revision INTEGER NOT NULL,
-                            current_revision INTEGER,
-                            status TEXT DEFAULT 'pending',
-                            outcome TEXT,
-                            attempt INTEGER DEFAULT 1,
-                            max_attempts INTEGER DEFAULT 3,
-                            step_status_json TEXT,
-                            started_at DATETIME,
-                            finished_at DATETIME,
-                            duration_ms INTEGER,
-                            error_class TEXT,
-                            error_message TEXT,
-                            created_at DATETIME,
-                            updated_at DATETIME,
-                            FOREIGN KEY(chat_id) REFERENCES chats(id)
-                        )
-                        """
-                    )
-                )
-                # Create indexes for query performance
-                conn.execute(
-                    text(
-                        "CREATE INDEX IF NOT EXISTS idx_postprocess_jobs_chat_id ON postprocess_jobs(chat_id)"
-                    )
-                )
-                conn.execute(
-                    text(
-                        "CREATE INDEX IF NOT EXISTS idx_postprocess_jobs_status ON postprocess_jobs(status)"
-                    )
-                )
-                conn.commit()
-
-        # Migration: Add cache token columns to cost_ledger and chat_cost_summary
-        if "cost_ledger" in inspector.get_table_names():
-            columns = [col["name"] for col in inspector.get_columns("cost_ledger")]
-            with self.engine.connect() as conn:
-                if "cache_write_tokens" not in columns:
-                    conn.execute(
-                        text(
-                            "ALTER TABLE cost_ledger ADD COLUMN cache_write_tokens INTEGER DEFAULT 0"
-                        )
-                    )
-                if "cache_read_tokens" not in columns:
-                    conn.execute(
-                        text(
-                            "ALTER TABLE cost_ledger ADD COLUMN cache_read_tokens INTEGER DEFAULT 0"
-                        )
-                    )
-                conn.commit()
-
-        if "chat_cost_summary" in inspector.get_table_names():
-            columns = [
-                col["name"] for col in inspector.get_columns("chat_cost_summary")
-            ]
-            with self.engine.connect() as conn:
-                if "total_cache_write_tokens" not in columns:
-                    conn.execute(
-                        text(
-                            "ALTER TABLE chat_cost_summary ADD COLUMN total_cache_write_tokens INTEGER DEFAULT 0"
-                        )
-                    )
-                if "total_cache_read_tokens" not in columns:
-                    conn.execute(
-                        text(
-                            "ALTER TABLE chat_cost_summary ADD COLUMN total_cache_read_tokens INTEGER DEFAULT 0"
-                        )
-                    )
-                conn.commit()
-
-        # Migration: Persist the latest rich context/usage payload per chat for UI display.
-        if "chats" in inspector.get_table_names():
-            columns = [col["name"] for col in inspector.get_columns("chats")]
-            with self.engine.connect() as conn:
-                if "context_usage" not in columns:
-                    conn.execute(
-                        text("ALTER TABLE chats ADD COLUMN context_usage JSON")
-                    )
-                conn.commit()
-
-        # Migration: Add projects table and chats.project_id FK.
-        # projects table is created by SQLModel.metadata.create_all() above,
-        # but chats.project_id needs an explicit ALTER on existing databases.
-        if "chats" in inspector.get_table_names():
-            columns = [col["name"] for col in inspector.get_columns("chats")]
-            if "project_id" not in columns:
-                with self.engine.connect() as conn:
-                    conn.execute(
-                        text(
-                            "ALTER TABLE chats ADD COLUMN project_id TEXT REFERENCES projects(id)"
-                        )
-                    )
-                    conn.execute(
-                        text(
-                            "CREATE INDEX IF NOT EXISTS ix_chats_project_id ON chats(project_id)"
-                        )
-                    )
-                    conn.commit()
-
-    def _session(self) -> Session:
-        """Create a new database session."""
-        return Session(self.engine)
-
-    def _migrate_static_config_from_db(self) -> None:
-        inspector = inspect(self.engine)
-        tables = set(inspector.get_table_names())
-        legacy_tables = {"user_preferences", "memory_config", "api_keys"} & tables
-        if not legacy_tables:
-            return
-
-        from suzent.core.user_config import UserConfigStore
-
-        store = UserConfigStore()
-        migrated = False
-        drop_tables: set[str] = set()
-
-        try:
-            with self.engine.connect() as conn:
-                did_migrate, can_drop = self._migrate_legacy_user_preferences(
-                    conn, store, legacy_tables
-                )
-                migrated |= did_migrate
-                if can_drop:
-                    drop_tables.add("user_preferences")
-
-                did_migrate, can_drop = self._migrate_legacy_memory_config(
-                    conn, store, legacy_tables
-                )
-                migrated |= did_migrate
-                if can_drop:
-                    drop_tables.add("memory_config")
-
-                did_migrate, can_drop = self._migrate_legacy_api_keys(
-                    conn, store, legacy_tables
-                )
-                migrated |= did_migrate
-                if can_drop:
-                    drop_tables.add("api_keys")
-
-                self._drop_legacy_static_tables(conn, drop_tables)
-                conn.commit()
-
-            if migrated:
-                from suzent.logger import logger
-
-                logger.info("Migrated legacy static config out of SQLite")
-        except Exception as exc:
-            from suzent.logger import logger
-
-            logger.warning(
-                "Failed to migrate legacy static config from SQLite: {}", exc
-            )
-
-    def _migrate_legacy_user_preferences(
-        self, conn: Any, store: Any, legacy_tables: set[str]
-    ) -> tuple[bool, bool]:
-        if "user_preferences" not in legacy_tables:
-            return False, False
-
-        row = (
-            conn.execute(
-                text(
-                    """
-                    SELECT model, agent, tools, memory_enabled,
-                           sandbox_enabled, sandbox_volumes
-                    FROM user_preferences WHERE id = 1
-                    """
-                )
-            )
-            .mappings()
-            .first()
-        )
-        if not row:
-            return False, True
-        if store.get_user_preferences():
-            from suzent.logger import logger
-
-            logger.warning(
-                "Keeping legacy user_preferences table because user config already exists"
-            )
-            return False, False
-
-        store.save_user_preferences(
-            {
-                "model": row["model"],
-                "agent": row["agent"],
-                "tools": self._decode_json_value(row["tools"]),
-                "memory_enabled": self._bool_or_none(row["memory_enabled"]),
-                "sandbox_enabled": self._bool_or_none(row["sandbox_enabled"]),
-                "sandbox_volumes": self._decode_json_value(row["sandbox_volumes"]),
-            }
-        )
-        return True, True
-
-    def _migrate_legacy_memory_config(
-        self, conn: Any, store: Any, legacy_tables: set[str]
-    ) -> tuple[bool, bool]:
-        if "memory_config" not in legacy_tables:
-            return False, False
-
-        row = (
-            conn.execute(
-                text(
-                    """
-                    SELECT embedding_model, extraction_model
-                    FROM memory_config WHERE id = 1
-                    """
-                )
-            )
-            .mappings()
-            .first()
-        )
-        if not row:
-            return False, True
-        if store.get_memory_config():
-            from suzent.logger import logger
-
-            logger.warning(
-                "Keeping legacy memory_config table because user config already exists"
-            )
-            return False, False
-
-        store.save_memory_config(
-            {
-                "embedding_model": row["embedding_model"],
-                "extraction_model": row["extraction_model"],
-            }
-        )
-        return True, True
-
-    def _migrate_legacy_api_keys(
-        self, conn: Any, store: Any, legacy_tables: set[str]
-    ) -> tuple[bool, bool]:
-        if "api_keys" not in legacy_tables:
-            return False, False
-
-        migrated = False
-        can_drop = True
-        rows = conn.execute(text("SELECT key, value FROM api_keys")).mappings().all()
-        for row in rows:
-            key = row["key"]
-            value = row["value"]
-            if not key or not value:
-                continue
-            if key.startswith("_"):
-                if store.get_config_blob(key) is None:
-                    store.save_config_blob(key, value)
-                    migrated = True
-                else:
-                    from suzent.logger import logger
-
-                    logger.warning(
-                        "Keeping legacy api_keys table because config blob '{}' already exists",
-                        key,
-                    )
-                    can_drop = False
-                continue
-            secret_migrated, secret_can_drop = self._migrate_secret_to_backend(
-                key, value
-            )
-            migrated |= secret_migrated
-            can_drop &= secret_can_drop
-        return migrated, can_drop
-
-    @staticmethod
-    def _drop_legacy_static_tables(conn: Any, legacy_tables: set[str]) -> None:
-        for table in legacy_tables:
-            conn.execute(text(f"DROP TABLE IF EXISTS {table}"))
-
-    @staticmethod
-    def _decode_json_value(value: Any) -> Any:
-        if value is None or not isinstance(value, str):
-            return value
-        try:
-            return json.loads(value)
-        except json.JSONDecodeError:
-            return value
-
-    @staticmethod
-    def _bool_or_none(value: Any) -> Optional[bool]:
-        if value is None:
-            return None
-        return bool(value)
-
-    @staticmethod
-    def _parse_config_datetime(value: Any) -> datetime:
-        if isinstance(value, datetime):
-            return value
-        if isinstance(value, str):
-            try:
-                return datetime.fromisoformat(value)
-            except ValueError:
-                pass
-        return datetime.now()
-
-    def _migrate_secret_to_backend(self, key: str, value: str) -> tuple[bool, bool]:
-        from suzent.core.secrets import get_secret_manager
-
-        secret_manager = get_secret_manager()
-        if secret_manager.has_backend_value(key):
-            from suzent.logger import logger
-
-            logger.warning(
-                "Keeping legacy api_keys table because secret '{}' already exists",
-                key,
-            )
-            return False, False
-
-        secret = self._decrypt_legacy_secret(key, value)
-        if secret is None:
-            return False, False
-        secret_manager.set_backend_only(key, secret)
-        return True, True
-
-    @staticmethod
-    def _decrypt_legacy_secret(key: str, value: str) -> Optional[str]:
-        try:
-            from cryptography.fernet import Fernet, InvalidToken
-            from suzent.config import DATA_DIR
-
-            raw_keys = []
-            if env_key := os.environ.get("SUZENT_SECRET_KEY"):
-                raw_keys.append(env_key.encode())
-            key_file = DATA_DIR / ".secret_key"
-            if key_file.exists():
-                raw_keys.append(key_file.read_bytes().strip())
-
-            for raw_key in raw_keys:
-                try:
-                    return Fernet(raw_key).decrypt(value.encode()).decode()
-                except InvalidToken:
-                    continue
-        except Exception:
-            pass
-
-        if ChatDatabase._looks_like_fernet_token(value):
-            from suzent.logger import logger
-
-            logger.warning(
-                "Keeping legacy api_keys table because secret '{}' could not be decrypted",
-                key,
-            )
-            return None
-        return value
-
-    @staticmethod
-    def _looks_like_fernet_token(value: str) -> bool:
-        try:
-            token = base64.urlsafe_b64decode(value.encode())
-        except Exception:
-            return False
-        return bool(token) and token[0] == 0x80
-
-    # -------------------------------------------------------------------------
-    # Project Operations
-    # -------------------------------------------------------------------------
-
-    DEFAULT_PROJECT_SLUG = "default"
-    DEFAULT_PROJECT_NAME = "Default"
-    SOCIAL_PROJECT_SLUG = "social"
-    SOCIAL_PROJECT_NAME = "Social"
-    SYSTEM_PROJECT_SLUGS = {"default", "social"}
-
-    # Platforms that route a chat into the ``social`` project by default.
-    # Internal platforms like ``subagent`` (lives with parent) and ``cron``
-    # (scheduled task runner) are deliberately excluded — subagents inherit
-    # their parent's project explicitly and cron chats stay in default.
-    SOCIAL_PLATFORMS = {"telegram", "slack", "discord", "wechat", "whatsapp"}
-
-    @classmethod
-    def _is_social_platform(cls, config: Optional[Dict[str, Any]]) -> bool:
-        if not config:
-            return False
-        platform = config.get("platform")
-        if not platform:
-            return False
-        return platform.lower() in cls.SOCIAL_PLATFORMS
-
-    def _migrate_legacy_session_dirs(self) -> None:
-        """Flatten legacy per-chat directories into the project root.
-
-        Handles three legacy shapes:
-
-        1. ``sandbox_data_path/sessions/{chat_id}/...`` (pre-project layout)
-        2. ``sandbox_data_path/projects/{slug}/chats/{chat_id}/...``
-           (intermediate layout — earlier migration step)
-        3. ``sandbox_data_path/shared/memory/sessions/{chat_id[:32]}/context.md``
-           (legacy memory context location)
-
-        After migration the project directory contains a flat layout::
-
-            projects/{slug}/
-              heartbeat.md
-              context.md
-              uploads/<files, prefixed with {chat_id[:8]}_ on collision>
-              images/<files, prefixed with {chat_id[:8]}_ on collision>
-
-        Each chat's files are merged into the project. On filename collision
-        within ``uploads/`` or ``images/`` the moved file is prefixed with the
-        short chat id so nothing is silently dropped. Single-file artifacts
-        (heartbeat.md, context.md) follow first-wins: the first chat's file
-        becomes the project's; subsequent chats' files are kept on disk under
-        a ``.from-chat-{chat_id[:8]}`` suffix so they can be reviewed.
-
-        Runs once per startup. Failures are logged but do not abort startup.
-        """
-        from suzent.logger import logger
-
-        try:
-            from suzent.config import CONFIG
-        except Exception:
-            return
-
-        import shutil
-
-        sandbox_root = Path(CONFIG.sandbox_data_path)
-        projects_root = sandbox_root / "projects"
-
-        # Phase 1: collect (chat_id, source_dir) pairs from legacy layouts.
-        sources: List[tuple[str, Path]] = []
-
-        # Legacy shape 1: sessions/{chat_id}/
-        sessions_root = sandbox_root / "sessions"
-        if sessions_root.exists() and sessions_root.is_dir():
-            for d in sessions_root.iterdir():
-                if d.is_dir():
-                    sources.append((d.name, d))
-
-        # Legacy shape 2: projects/{slug}/chats/{chat_id}/
-        if projects_root.exists():
-            for project_dir in projects_root.iterdir():
-                chats_dir = project_dir / "chats"
-                if chats_dir.exists() and chats_dir.is_dir():
-                    for d in chats_dir.iterdir():
-                        if d.is_dir():
-                            sources.append((d.name, d))
-
-        moved_chats = 0
-        for chat_id, source_dir in sources:
-            try:
-                if self._flatten_chat_dir_into_project(chat_id, source_dir):
-                    moved_chats += 1
-            except Exception as e:
-                logger.warning("Failed to flatten chat dir {}: {}", source_dir, e)
-
-        if moved_chats:
-            logger.info(
-                "Flattened {} legacy chat dir(s) into project roots", moved_chats
-            )
-
-        # Legacy shape 3: shared/memory/sessions/{chat_id[:32]}/context.md
-        mem_sessions = sandbox_root / "shared" / "memory" / "sessions"
-        if mem_sessions.exists() and mem_sessions.is_dir():
-            moved_ctx = 0
-            for d in mem_sessions.iterdir():
-                if not d.is_dir():
-                    continue
-                ctx = d / "context.md"
-                if not ctx.exists():
-                    continue
-                # d.name is chat_id[:32] — find a chat whose id starts with it
-                short_id = d.name
-                chat_id = self._find_chat_id_by_prefix(short_id)
-                if not chat_id:
-                    logger.debug(
-                        "No chat matches legacy memory dir {}; leaving in place",
-                        short_id,
-                    )
-                    continue
-                slug = self.get_chat_project_slug(chat_id)
-                project_dir = projects_root / slug
-                project_dir.mkdir(parents=True, exist_ok=True)
-                dest = project_dir / "context.md"
-                if dest.exists():
-                    # First-wins; keep this chat's context under a backup name
-                    backup = project_dir / f"context.from-chat-{chat_id[:8]}.md"
-                    if not backup.exists():
-                        shutil.move(str(ctx), str(backup))
-                else:
-                    shutil.move(str(ctx), str(dest))
-                # Clean up the empty legacy dir
-                try:
-                    if not any(d.iterdir()):
-                        d.rmdir()
-                except Exception:
-                    pass
-                moved_ctx += 1
-            if moved_ctx:
-                logger.info("Migrated {} legacy memory context file(s)", moved_ctx)
-            # Drop the empty sessions root
-            try:
-                if mem_sessions.exists() and not any(mem_sessions.iterdir()):
-                    mem_sessions.rmdir()
-            except Exception:
-                pass
-
-        # Drop empty top-level legacy dirs
-        for empty_root in (sessions_root,):
-            try:
-                if empty_root.exists() and not any(empty_root.iterdir()):
-                    empty_root.rmdir()
-            except Exception:
-                pass
-
-        # Drop empty projects/{slug}/chats/ directories
-        if projects_root.exists():
-            for project_dir in projects_root.iterdir():
-                chats_dir = project_dir / "chats"
-                try:
-                    if chats_dir.exists() and not any(chats_dir.iterdir()):
-                        chats_dir.rmdir()
-                except Exception:
-                    pass
-
-    def _flatten_chat_dir_into_project(self, chat_id: str, source_dir: Path) -> bool:
-        """Move the contents of a legacy chat directory into its project root.
-
-        Returns True if at least one file was moved. See _migrate_legacy_session_dirs
-        for the collision/merge rules.
-        """
-        import shutil
-        from suzent.logger import logger
-
-        if not source_dir.exists():
-            return False
-
-        slug = self.get_chat_project_slug(chat_id)
-        from suzent.config import CONFIG
-
-        project_dir = Path(CONFIG.sandbox_data_path) / "projects" / slug
-        project_dir.mkdir(parents=True, exist_ok=True)
-
-        short = chat_id[:8]
-        moved_any = False
-
-        for entry in list(source_dir.iterdir()):
-            try:
-                if entry.is_dir() and entry.name in ("uploads", "images"):
-                    # Merge contents file-by-file, prefixing collisions with chat id
-                    dest_dir = project_dir / entry.name
-                    dest_dir.mkdir(parents=True, exist_ok=True)
-                    for f in list(entry.iterdir()):
-                        target = dest_dir / f.name
-                        if target.exists():
-                            target = dest_dir / f"{short}_{f.name}"
-                        if target.exists():
-                            # Even the prefixed name exists; skip to avoid clobber
-                            logger.debug(
-                                "Skipping {} for chat {}: collision even after prefix",
-                                f,
-                                short,
-                            )
-                            continue
-                        shutil.move(str(f), str(target))
-                        moved_any = True
-                    # Remove the empty source subdir
-                    try:
-                        if not any(entry.iterdir()):
-                            entry.rmdir()
-                    except Exception:
-                        pass
-                elif entry.is_file() and entry.name in ("heartbeat.md", "context.md"):
-                    # Single-file artifacts: first-wins, keep extras as backup
-                    dest = project_dir / entry.name
-                    if dest.exists():
-                        backup = (
-                            project_dir
-                            / f"{entry.stem}.from-chat-{short}{entry.suffix}"
-                        )
-                        if not backup.exists():
-                            shutil.move(str(entry), str(backup))
-                            moved_any = True
-                    else:
-                        shutil.move(str(entry), str(dest))
-                        moved_any = True
-                else:
-                    # Unknown file/dir — move with prefix on collision
-                    target = project_dir / entry.name
-                    if target.exists():
-                        target = project_dir / f"{short}_{entry.name}"
-                    if not target.exists():
-                        shutil.move(str(entry), str(target))
-                        moved_any = True
-            except Exception as e:
-                logger.warning("Failed migrating {} from chat {}: {}", entry, short, e)
-
-        # Remove the now-empty source dir
-        try:
-            if source_dir.exists() and not any(source_dir.iterdir()):
-                source_dir.rmdir()
-        except Exception:
-            pass
-
-        return moved_any
-
-    def _find_chat_id_by_prefix(self, prefix: str) -> Optional[str]:
-        """Resolve a chat id from its first 32 chars (used by legacy memory dirs)."""
-        with self._session() as session:
-            for chat in session.exec(
-                select(ChatModel).where(ChatModel.id.startswith(prefix))
-            ).all():
-                return chat.id
-        return None
-
-    def _ensure_default_project(self) -> None:
-        """Create system projects (default + social) and backfill chats.
-
-        - ``default``: catches chats with no platform set.
-        - ``social``: catches chats with ``config.platform`` set (Telegram, etc.).
-
-        Both system projects are created on first launch and any unassigned
-        chats are routed to the appropriate one. Existing chats already in
-        ``default`` that have a ``config.platform`` set are also moved to
-        ``social`` (one-time migration).
-        """
-        from suzent.logger import logger
-
-        with self._session() as session:
-            # System projects
-            default = session.exec(
-                select(ProjectModel).where(
-                    ProjectModel.slug == self.DEFAULT_PROJECT_SLUG
-                )
-            ).first()
-            if not default:
-                default = ProjectModel(
-                    name=self.DEFAULT_PROJECT_NAME,
-                    slug=self.DEFAULT_PROJECT_SLUG,
-                )
-                session.add(default)
-                session.flush()
-                logger.info(
-                    "Created default project (slug='{}')", self.DEFAULT_PROJECT_SLUG
-                )
-
-            social = session.exec(
-                select(ProjectModel).where(
-                    ProjectModel.slug == self.SOCIAL_PROJECT_SLUG
-                )
-            ).first()
-            if not social:
-                social = ProjectModel(
-                    name=self.SOCIAL_PROJECT_NAME,
-                    slug=self.SOCIAL_PROJECT_SLUG,
-                )
-                session.add(social)
-                session.flush()
-                logger.info(
-                    "Created social project (slug='{}')", self.SOCIAL_PROJECT_SLUG
-                )
-
-            # Backfill unassigned chats: route by config.platform presence
-            unassigned = session.exec(
-                select(ChatModel).where(ChatModel.project_id == None)  # noqa: E711
-            ).all()
-            backfilled_default = 0
-            backfilled_social = 0
-            for chat in unassigned:
-                if self._is_social_platform(chat.config):
-                    chat.project_id = social.id
-                    backfilled_social += 1
-                else:
-                    chat.project_id = default.id
-                    backfilled_default += 1
-            if backfilled_default:
-                logger.info(
-                    "Backfilled {} chat(s) to default project", backfilled_default
-                )
-            if backfilled_social:
-                logger.info(
-                    "Backfilled {} chat(s) to social project", backfilled_social
-                )
-
-            # One-time migration: chats already in default that have a social
-            # platform tag were placed there before the social project existed.
-            # Subagent/cron chats are excluded — they stay where they were.
-            misplaced = session.exec(
-                select(ChatModel).where(ChatModel.project_id == default.id)
-            ).all()
-            moved_to_social = 0
-            for chat in misplaced:
-                if self._is_social_platform(chat.config):
-                    chat.project_id = social.id
-                    moved_to_social += 1
-
-            # One-time rescue: subagent/cron chats that landed in social in a
-            # prior migration step need to move out. Subagents follow their
-            # parent's project (falling back to default if parent is unknown);
-            # cron chats go to default.
-            misclassified = session.exec(
-                select(ChatModel).where(ChatModel.project_id == social.id)
-            ).all()
-            rescued_subagent = 0
-            rescued_cron = 0
-            for chat in misclassified:
-                if not chat.config:
-                    continue
-                platform = (chat.config.get("platform") or "").lower()
-                if platform == "subagent":
-                    parent_id = chat.config.get("parent_chat_id")
-                    target_project_id = default.id
-                    if parent_id:
-                        parent = session.get(ChatModel, parent_id)
-                        if parent and parent.project_id:
-                            target_project_id = parent.project_id
-                    chat.project_id = target_project_id
-                    rescued_subagent += 1
-                elif platform == "cron":
-                    chat.project_id = default.id
-                    rescued_cron += 1
-            if rescued_subagent:
-                logger.info(
-                    "Moved {} subagent chat(s) out of social into parent project",
-                    rescued_subagent,
-                )
-            if rescued_cron:
-                logger.info(
-                    "Moved {} cron chat(s) out of social into default", rescued_cron
-                )
-            if moved_to_social:
-                logger.info(
-                    "Migrated {} platform-tagged chat(s) from default to social",
-                    moved_to_social,
-                )
-
-            session.commit()
-
-    def create_project(self, name: str, slug: str) -> str:
-        """Create a new project and return its id."""
-        project = ProjectModel(name=name, slug=slug)
-        with self._session() as session:
-            session.add(project)
-            session.commit()
-            session.refresh(project)
-            return project.id
-
-    def get_project(self, project_id: str) -> Optional[ProjectModel]:
-        """Return a project by id, or None if not found."""
-        with self._session() as session:
-            return session.get(ProjectModel, project_id)
-
-    def get_project_by_slug(self, slug: str) -> Optional[ProjectModel]:
-        """Return a project by slug, or None if not found."""
-        with self._session() as session:
-            return session.exec(
-                select(ProjectModel).where(ProjectModel.slug == slug)
-            ).first()
-
-    def list_projects(self, include_archived: bool = False) -> List[ProjectModel]:
-        """Return all projects ordered by creation time."""
-        with self._session() as session:
-            stmt = select(ProjectModel)
-            if not include_archived:
-                stmt = stmt.where(ProjectModel.archived == False)  # noqa: E712
-            stmt = stmt.order_by(ProjectModel.created_at)
-            return list(session.exec(stmt).all())
-
-    @staticmethod
-    def _is_subagent_child(chat: ChatModel, parent_chat_ids: set[str]) -> bool:
-        """Return whether a chat is a direct subagent child of one of the parents."""
-        config = chat.config or {}
-        return (
-            config.get("platform") == "subagent"
-            and config.get("parent_chat_id") in parent_chat_ids
-        )
-
-    def get_subagent_chat_ids_for_parent_chat(self, chat_id: str) -> List[str]:
-        """Return direct subagent chat ids whose parent is the given chat."""
-        with self._session() as session:
-            chats = session.exec(select(ChatModel)).all()
-            return [
-                chat.id for chat in chats if self._is_subagent_child(chat, {chat_id})
-            ]
-
-    def list_subagent_task_records(
-        self, parent_chat_id: Optional[str] = None
-    ) -> List[dict]:
-        """Return persisted subagent task records reconstructed from chat rows.
-
-        Subagent runtime tasks are in memory, but their isolated chat rows are
-        persisted. This lets the UI show historical subagents after restart and
-        when the currently selected chat is the subagent chat itself.
-        """
-        with self._session() as session:
-            chats = session.exec(select(ChatModel)).all()
-
-        records = []
-        for chat in chats:
-            config = chat.config or {}
-            if config.get("platform") != "subagent":
-                continue
-
-            task_id = config.get("subagent_task_id")
-            if not task_id and chat.id.startswith("subagent-"):
-                task_id = chat.id[len("subagent-") :]
-
-            parent_id = config.get("parent_chat_id")
-
-            if parent_chat_id and parent_chat_id not in {parent_id, chat.id}:
-                continue
-
-            description = chat.title
-            prefix = "Sub-agent:"
-            if description.startswith(prefix):
-                description = description[len(prefix) :].strip()
-
-            records.append(
-                {
-                    "task_id": task_id,
-                    "parent_chat_id": parent_id or "",
-                    "chat_id": chat.id,
-                    "description": description,
-                    "tools_allowed": config.get("tools") or [],
-                    "status": "completed",
-                    "result_summary": None,
-                    "error": None,
-                    "model_override": config.get("model"),
-                    "started_at": chat.created_at.isoformat()
-                    if chat.created_at
-                    else None,
-                    "finished_at": chat.last_result_at.isoformat()
-                    if chat.last_result_at
-                    else chat.updated_at.isoformat()
-                    if chat.updated_at
-                    else None,
-                    "inherit_context": config.get("inherit_context", False),
-                    "isolation": config.get("isolation", "none"),
-                    "worktree_path": config.get("worktree_path"),
-                    "worktree_branch": config.get("worktree_branch"),
-                }
-            )
-
-        return sorted(
-            records,
-            key=lambda record: record.get("finished_at")
-            or record.get("started_at")
-            or "",
-            reverse=True,
-        )
-
-    def link_chat_to_project(self, chat_id: str, project_id: str) -> bool:
-        """Move a chat and its direct subagent chats to a different project."""
-        with self._session() as session:
-            chat = session.get(ChatModel, chat_id)
-            if not chat:
-                return False
-            chats = session.exec(select(ChatModel)).all()
-            moved_chats = [
-                candidate
-                for candidate in chats
-                if candidate.id == chat_id
-                or self._is_subagent_child(candidate, {chat_id})
-            ]
-            for moved_chat in moved_chats:
-                moved_chat.project_id = project_id
-                session.add(moved_chat)
-            session.commit()
-            return True
-
-    def move_all_chats(self, from_project_id: str, to_project_id: str) -> int:
-        """Reassign every chat from one project plus their subagents."""
-        with self._session() as session:
-            project_chats = session.exec(
-                select(ChatModel).where(ChatModel.project_id == from_project_id)
-            ).all()
-            project_chat_ids = {chat.id for chat in project_chats}
-            all_chats = session.exec(select(ChatModel)).all()
-            moved_chats = [
-                chat
-                for chat in all_chats
-                if chat.id in project_chat_ids
-                or self._is_subagent_child(chat, project_chat_ids)
-            ]
-            for chat in moved_chats:
-                chat.project_id = to_project_id
-                session.add(chat)
-            session.commit()
-            return len(moved_chats)
-
-    def get_chat_ids_in_project(self, project_id: str) -> List[str]:
-        """Return chat ids assigned to a project plus direct subagents of those chats."""
-        with self._session() as session:
-            project_chats = session.exec(
-                select(ChatModel).where(ChatModel.project_id == project_id)
-            ).all()
-            project_chat_ids = {chat.id for chat in project_chats}
-            all_chats = session.exec(select(ChatModel)).all()
-            return [
-                chat.id
-                for chat in all_chats
-                if chat.id in project_chat_ids
-                or self._is_subagent_child(chat, project_chat_ids)
-            ]
-
-    def update_project(
-        self,
-        project_id: str,
-        name: Optional[str] = None,
-        archived: Optional[bool] = None,
-    ) -> Optional[ProjectModel]:
-        """Update a project's name and/or archived flag."""
-        with self._session() as session:
-            project = session.get(ProjectModel, project_id)
-            if not project:
-                return None
-            if name is not None:
-                project.name = name
-            if archived is not None:
-                project.archived = archived
-            session.add(project)
-            session.commit()
-            session.refresh(project)
-            return project
-
-    def delete_project(self, project_id: str) -> tuple[bool, str]:
-        """Delete a project. Refuses to delete system projects or non-empty projects.
-
-        Returns ``(success, error_message)``. On success error_message is empty.
-        """
-        with self._session() as session:
-            project = session.get(ProjectModel, project_id)
-            if not project:
-                return False, "Project not found"
-            if project.slug in self.SYSTEM_PROJECT_SLUGS:
-                return False, f"Cannot delete system project '{project.slug}'"
-            chat_count = session.exec(
-                select(func.count())
-                .select_from(ChatModel)
-                .where(ChatModel.project_id == project_id)
-            ).one()
-            if chat_count > 0:
-                return (
-                    False,
-                    f"Project has {chat_count} chat(s); move them out first",
-                )
-            session.delete(project)
-            session.commit()
-            return True, ""
-
-    def get_chat_project_id(self, chat_id: str) -> Optional[str]:
-        """Return the project_id for a chat, or None if not found."""
-        with self._session() as session:
-            chat = session.get(ChatModel, chat_id)
-            return chat.project_id if chat else None
-
-    def get_chat_project_slug(self, chat_id: str) -> str:
-        """Return the project slug for a chat, falling back to the default project slug.
-
-        Used by path resolution to map a chat to its on-disk project directory.
-        """
-        project_id = self.get_chat_project_id(chat_id)
-        if project_id:
-            project = self.get_project(project_id)
-            if project:
-                return project.slug
-        return self.DEFAULT_PROJECT_SLUG
-
-    def get_project_dir(self, chat_id: str) -> Path:
-        """Host path for the project a chat belongs to.
-
-        This directory is the agent's cwd and contains all project-scoped state:
-        heartbeat.md, context.md, plan.md, uploads/, images/, plus any files
-        the agent creates. Shared across all chats in the project.
-        """
-        from suzent.config import CONFIG
-
-        slug = self.get_chat_project_slug(chat_id)
-        return Path(CONFIG.sandbox_data_path) / "projects" / slug
-
-    def find_heartbeat_enabled_chat_in_project(
-        self,
-        project_id: str,
-        exclude_chat_id: Optional[str] = None,
-        exclude_chat_ids: Optional[set[str]] = None,
-    ) -> Optional["ChatModel"]:
-        """Return the chat in this project that already has heartbeat enabled.
-
-        Used to enforce one-heartbeat-per-project. Returns None if no other chat
-        in the project has heartbeat enabled.
-        """
-        excluded = set(exclude_chat_ids or set())
-        if exclude_chat_id is not None:
-            excluded.add(exclude_chat_id)
-
-        with self._session() as session:
-            stmt = select(ChatModel).where(ChatModel.project_id == project_id)
-            if excluded:
-                stmt = stmt.where(ChatModel.id.not_in(excluded))
-            for chat in session.exec(stmt).all():
-                if chat.config and chat.config.get("heartbeat_enabled"):
-                    return chat
-            return None
-
-    # -------------------------------------------------------------------------
-    # Chat Operations
-    # -------------------------------------------------------------------------
-
-    def create_chat(
-        self,
-        title: str,
-        config: Dict[str, Any],
-        messages: List[Dict[str, Any]] = None,
-        agent_state: bytes = None,
-        chat_id: str = None,
-        working_directory: str = None,
-        context_usage: Dict[str, Any] = None,
-        project_id: str = None,
-    ) -> str:
-        """Create a new chat and return its ID."""
-        now = datetime.now()
-        chat_id = chat_id or str(uuid.uuid4())
-
-        # Default to the default project if none specified
-        if project_id is None:
-            default = self.get_project_by_slug(self.DEFAULT_PROJECT_SLUG)
-            if default:
-                project_id = default.id
-
-        chat = ChatModel(
-            id=chat_id,
-            title=title,
-            created_at=now,
-            updated_at=now,
-            config=config,
-            messages=messages or [],
-            context_usage=context_usage or {},
-            agent_state=agent_state,
-            working_directory=working_directory,
-            project_id=project_id,
-        )
-
-        with self._session() as session:
-            session.add(chat)
-            session.commit()
-
-        return chat_id
-
-    def get_chat(self, chat_id: str) -> Optional[ChatModel]:
-        """Get a specific chat by ID."""
-        with self._session() as session:
-            return session.get(ChatModel, chat_id)
-
-    def update_chat(
-        self,
-        chat_id: str,
-        title: str = None,
-        config: Dict[str, Any] = None,
-        messages: List[Dict[str, Any]] = None,
-        agent_state: bytes = None,
-        working_directory: str = None,
-        context_usage: Dict[str, Any] = None,
-    ) -> bool:
-        """Update an existing chat."""
-        with self._session() as session:
-            chat = session.get(ChatModel, chat_id)
-            if not chat:
-                return False
-
-            should_update_timestamp = False
-
-            if title is not None and title != chat.title:
-                chat.title = title
-                should_update_timestamp = True
-
-            if config is not None:
-                chat.config = config
-                flag_modified(chat, "config")
-
-            if messages is not None:
-                chat.messages = messages
-                should_update_timestamp = True
-
-            if context_usage is not None:
-                chat.context_usage = context_usage
-                flag_modified(chat, "context_usage")
-
-            if agent_state is not None:
-                chat.agent_state = agent_state
-                should_update_timestamp = True
-
-            if working_directory is not None:
-                chat.working_directory = working_directory
-
-            if should_update_timestamp:
-                chat.updated_at = datetime.now()
-
-            session.add(chat)
-            session.commit()
-            return True
-
-    def commit_snapshot_state(self, chat_id: str, agent_state: bytes) -> Optional[int]:
-        """Commit fast snapshot state and increment revision atomically.
-
-        Returns:
-            The new state revision if the chat exists, otherwise None.
-        """
-        now = datetime.now()
-        with self._session() as session:
-            chat = session.get(ChatModel, chat_id)
-            if not chat:
-                _postprocess_metrics.snapshot_failed += 1
-                return None
-
-            next_revision = (chat.state_revision or 0) + 1
-            chat.agent_state = agent_state
-            chat.state_revision = next_revision
-            chat.state_stage = "snapshot"
-            chat.state_updated_at = now
-            chat.updated_at = now
-
-            session.add(chat)
-            session.commit()
-
-        _postprocess_metrics.snapshot_committed += 1
-        return next_revision
-
-    def finalize_state_if_revision_matches(
-        self,
-        chat_id: str,
-        expected_revision: int,
-        agent_state: bytes,
-        messages: Optional[List[Dict[str, Any]]] = None,
-        update_lifecycle: bool = True,
-    ) -> bool:
-        """Finalize chat state only when revision matches expected value.
-
-        Returns:
-            True when finalized commit succeeds, False when stale/not found.
-        """
-        now = datetime.now()
-        with self._session() as session:
-            chat = session.get(ChatModel, chat_id)
-            if not chat:
-                return False
-
-            current_revision = chat.state_revision or 0
-            if current_revision != expected_revision:
-                return False
-
-            chat.agent_state = agent_state
-            if messages is not None:
-                chat.messages = messages
-            chat.finalized_revision = expected_revision
-            chat.state_stage = "finalized"
-            chat.state_updated_at = now
-            chat.updated_at = now
-
-            if update_lifecycle:
-                chat.last_active_at = now
-                chat.turn_count = (chat.turn_count or 0) + 1
-                config = dict(chat.config or {})
-                config["unread_count"] = config.get("unread_count", 0) + 1
-                chat.config = config
-                flag_modified(chat, "config")
-
-            session.add(chat)
-            session.commit()
-            return True
-
-    def delete_chat(self, chat_id: str, cascade_subagents: bool = False) -> bool:
-        """Delete a chat by ID. Optionally cascade-delete subagent children."""
-        with self._session() as session:
-            chat = session.get(ChatModel, chat_id)
-            if not chat:
-                return False
-
-            if cascade_subagents:
-                all_chats = session.exec(select(ChatModel)).all()
-                for c in all_chats:
-                    if self._is_subagent_child(c, {chat_id}):
-                        session.delete(c)
-
-            session.delete(chat)
-            session.commit()
-            return True
-
-    def set_last_result_at(self, chat_id: str) -> None:
-        """Mark that a background executor just completed a turn for this chat."""
-        with self._session() as session:
-            chat = session.get(ChatModel, chat_id)
-            if chat:
-                chat.last_result_at = datetime.now(timezone.utc)
-                session.add(chat)
-                session.commit()
-
-    def mark_chat_read(self, chat_id: str) -> None:
-        """Reset unread count to zero when the user views a chat."""
-        with self._session() as session:
-            chat = session.get(ChatModel, chat_id)
-            if chat:
-                config = dict(chat.config or {})
-                config["unread_count"] = 0
-                chat.config = config
-                flag_modified(chat, "config")
-                session.add(chat)
-                session.commit()
-
-    def list_chats(
-        self,
-        limit: int = 50,
-        offset: int = 0,
-        search: str = None,
-        platform: str = None,
-        project_id: str = None,
-    ) -> List[ChatSummaryModel]:
-        """List chat summaries ordered by last updated.
-
-        Args:
-            project_id: Optional filter — only return chats in this project.
-            platform: Optional ``"social"`` / ``"personal"`` filter on the legacy
-                ``config.platform`` field. Retained for backward compatibility;
-                new code should prefer ``project_id``.
-        """
-        with self._session() as session:
-            statement = (
-                select(
-                    ChatModel.id,
-                    ChatModel.title,
-                    ChatModel.created_at,
-                    ChatModel.updated_at,
-                    ChatModel.config,
-                    ChatModel.messages,
-                    ChatModel.last_result_at,
-                    ChatModel.project_id,
-                )
-                .order_by(ChatModel.updated_at.desc())
-                .offset(offset)
-                .limit(limit)
-            )
-
-            if search:
-                statement = statement.where(
-                    or_(
-                        ChatModel.title.contains(search),
-                        _messages_search_filter(search),
-                    )
-                )
-
-            if project_id is not None:
-                statement = statement.where(ChatModel.project_id == project_id)
-
-            if platform == "social":
-                statement = statement.where(
-                    text("json_extract(config, '$.platform') IS NOT NULL")
-                )
-            elif platform == "personal":
-                statement = statement.where(
-                    text("json_extract(config, '$.platform') IS NULL")
-                )
-
-            # Build a slug/name lookup so each summary row can carry project info
-            # without a per-row query.
-            project_lookup: Dict[str, ProjectModel] = {
-                p.id: p for p in session.exec(select(ProjectModel)).all()
-            }
-
-            results = []
-            import re
-
-            for row in session.exec(statement).all():
-                (
-                    chat_id,
-                    title,
-                    created_at,
-                    updated_at,
-                    config,
-                    messages,
-                    last_result_at,
-                    row_project_id,
-                ) = row
-                messages = messages or []
-                last_message = None
-                if messages:
-                    last_msg = messages[-1]
-                    if isinstance(last_msg, dict):
-                        content = last_msg.get("content") or ""
-                    else:
-                        content = str(last_msg)
-
-                    if isinstance(content, list):
-                        # Extract text from list of dicts (e.g. OpenAI vision)
-                        text_parts = []
-                        for part in content:
-                            if isinstance(part, dict) and part.get("type") == "text":
-                                text_parts.append(part.get("text", ""))
-                            elif isinstance(part, str):
-                                text_parts.append(part)
-                        content = " ".join(text_parts)
-                    elif not isinstance(content, str):
-                        content = str(content)
-
-                    # 1. completely eradicate any <details> tool blocks (including inner text)
-                    clean_content = re.sub(
-                        r"<details\b[^>]*>.*?</details>", "", content, flags=re.DOTALL
-                    )
-                    # 2. Strip remaining raw HTML tags
-                    clean_content = re.sub(r"<[^>]+>", "", clean_content).strip()
-
-                    last_message = clean_content[:100]
-                    if len(clean_content) > 100:
-                        last_message += "..."
-
-                config = config or {}
-                # Count only assistant messages that carry visible text.
-                # Pure tool-call turns have role="assistant" but no text content,
-                # so we exclude them to avoid inflating the unread badge.
-                visible_count = sum(
-                    1
-                    for m in messages
-                    if isinstance(m, dict)
-                    and m.get("role") == "assistant"
-                    and isinstance(m.get("content"), str)
-                    and m["content"].strip()
-                )
-                project = project_lookup.get(row_project_id) if row_project_id else None
-                results.append(
-                    ChatSummaryModel(
-                        id=chat_id,
-                        title=title,
-                        createdAt=created_at.isoformat(),
-                        updatedAt=updated_at.isoformat(),
-                        messageCount=visible_count,
-                        lastMessage=last_message,
-                        platform=config.get("platform"),
-                        heartbeatEnabled=config.get("heartbeat_enabled", False),
-                        lastResultAt=last_result_at.isoformat()
-                        if last_result_at
-                        else None,
-                        unreadCount=config.get("unread_count", 0),
-                        projectId=row_project_id,
-                        projectSlug=project.slug if project else None,
-                        projectName=project.name if project else None,
-                        parentChatId=config.get("parent_chat_id"),
-                    )
-                )
-
-            return results
-
-    def get_chat_count(
-        self, search: str = None, platform: str = None, project_id: str = None
-    ) -> int:
-        """Get total number of chats."""
-        with self._session() as session:
-            statement = select(func.count()).select_from(ChatModel)
-            if search:
-                statement = statement.where(
-                    or_(
-                        ChatModel.title.contains(search),
-                        _messages_search_filter(search),
-                    )
-                )
-            if project_id is not None:
-                statement = statement.where(ChatModel.project_id == project_id)
-            if platform == "social":
-                statement = statement.where(
-                    text("json_extract(config, '$.platform') IS NOT NULL")
-                )
-            elif platform == "personal":
-                statement = statement.where(
-                    text("json_extract(config, '$.platform') IS NULL")
-                )
-            return session.exec(statement).one()
-
-    def get_chat_kind_counts(
-        self, search: str = None, platform: str = None, project_id: str = None
-    ) -> Dict[str, int]:
-        """Get chat totals for the sidebar kind tabs."""
-
-        def apply_common_filters(statement):
-            if search:
-                statement = statement.where(
-                    or_(
-                        ChatModel.title.contains(search),
-                        _messages_search_filter(search),
-                    )
-                )
-            if project_id is not None:
-                statement = statement.where(ChatModel.project_id == project_id)
-            if platform == "social":
-                statement = statement.where(
-                    text("json_extract(config, '$.platform') IS NOT NULL")
-                )
-            elif platform == "personal":
-                statement = statement.where(
-                    text("json_extract(config, '$.platform') IS NULL")
-                )
-            return statement
-
-        def count(extra_filter=None) -> int:
-            statement = select(func.count()).select_from(ChatModel)
-            statement = apply_common_filters(statement)
-            if extra_filter is not None:
-                statement = statement.where(extra_filter)
-            return session.exec(statement).one()
-
-        with self._session() as session:
-            scheduled_filter = text(
-                "lower(json_extract(config, '$.platform')) = 'cron'"
-            )
-            you_filter = text(
-                "(json_extract(config, '$.platform') IS NULL "
-                "OR lower(json_extract(config, '$.platform')) != 'cron')"
-            )
-            total = count()
-            scheduled = count(scheduled_filter)
-            return {
-                "you": count(you_filter),
-                "scheduled": scheduled,
-                "all": total,
-            }
-
-    def count_chats_in_project(self, project_id: str) -> int:
-        """Return the number of chats currently in this project."""
-        with self._session() as session:
-            return session.exec(
-                select(func.count())
-                .select_from(ChatModel)
-                .where(ChatModel.project_id == project_id)
-            ).one()
-
-    def reassign_plan_chat(self, old_chat_id: str, new_chat_id: str) -> int:
-        """Reassign all plans from one chat_id to another."""
-        if old_chat_id == new_chat_id:
-            return 0
-
-        with self._session() as session:
-            statement = select(PlanModel).where(PlanModel.chat_id == old_chat_id)
-            plans = session.exec(statement).all()
-
-            for plan in plans:
-                plan.chat_id = new_chat_id
-                plan.updated_at = datetime.now()
-                session.add(plan)
-
-            session.commit()
-            return len(plans)
-
-    def get_active_heartbeats(self) -> List[ChatModel]:
-        """Get all chats that have heartbeat enabled in their config."""
-        with self._session() as session:
-            statement = select(
-                ChatModel.id,
-                ChatModel.title,
-                ChatModel.created_at,
-                ChatModel.updated_at,
-                ChatModel.config,
-            )
-            chats = []
-            for chat_id, title, created_at, updated_at, config in session.exec(
-                statement
-            ).all():
-                config = config or {}
-                if config.get("heartbeat_enabled") is True:
-                    chats.append(
-                        ChatModel(
-                            id=chat_id,
-                            title=title,
-                            created_at=created_at,
-                            updated_at=updated_at,
-                            config=config,
-                        )
-                    )
-            return chats
-
-    # -------------------------------------------------------------------------
-    # PostProcess Job Operations
-    # -------------------------------------------------------------------------
-
-    def create_postprocess_job(
-        self,
-        job_id: str,
-        chat_id: str,
-        assigned_revision: int,
-        max_attempts: int = 3,
-    ) -> Optional[Dict[str, Any]]:
-        """Create a new postprocess job record.
-
-        Args:
-            job_id: UUID-based unique job identifier
-            chat_id: Chat ID for this job
-            assigned_revision: Revision from Phase A snapshot
-            max_attempts: Max retry attempts (default: 3)
-
-        Returns:
-            Dict with job data if successful, None otherwise.
-        """
-        now = datetime.now()
-        with self._session() as session:
-            # Verify chat exists
-            chat = session.get(ChatModel, chat_id)
-            if not chat:
-                return None
-
-            job = PostprocessJobModel(
-                job_id=job_id,
-                chat_id=chat_id,
-                assigned_revision=assigned_revision,
-                status=PostProcessStatus.PENDING,
-                attempt=1,
-                max_attempts=max_attempts,
-                created_at=now,
-                updated_at=now,
-            )
-            session.add(job)
-            session.commit()
-
-            # Increment metrics
-            _postprocess_metrics.job_started += 1
-
-            # Return dict representation while still in session context
-            return {
-                "id": job.id,
-                "job_id": job.job_id,
-                "chat_id": job.chat_id,
-                "assigned_revision": job.assigned_revision,
-                "status": job.status,
-                "attempt": job.attempt,
-                "max_attempts": job.max_attempts,
-            }
-
-    def start_postprocess_job(self, job_id: str) -> bool:
-        """Mark a postprocess job as running."""
-        now = datetime.now()
-        with self._session() as session:
-            statement = select(PostprocessJobModel).where(
-                PostprocessJobModel.job_id == job_id
-            )
-            job = session.exec(statement).first()
-            if not job:
-                return False
-
-            job.status = PostProcessStatus.RUNNING
-            job.started_at = now
-            job.updated_at = now
-            session.add(job)
-            session.commit()
-            return True
-
-    def update_job_step_status(
-        self,
-        job_id: str,
-        step: str,
-        status: str,
-        error: str = None,
-        duration_ms: int = None,
-    ) -> bool:
-        """Update status of a specific step within a postprocess job.
-
-        Args:
-            job_id: Job identifier
-            step: Step name (B1_transcript, B2_memory, etc.)
-            status: Step status (pending, running, success, failed)
-            error: Error message if failed
-            duration_ms: Duration of the step in milliseconds
-
-        Returns:
-            True if updated successfully, False if job not found.
-        """
-        now = datetime.now()
-        with self._session() as session:
-            statement = select(PostprocessJobModel).where(
-                PostprocessJobModel.job_id == job_id
-            )
-            job = session.exec(statement).first()
-            if not job:
-                return False
-
-            # Parse or initialize step_status_json
-            try:
-                step_status = json.loads(job.step_status_json or "{}")
-            except (json.JSONDecodeError, TypeError):
-                step_status = {}
-
-            # Update step status
-            step_status[step] = {
-                "status": status,
-                "error": error,
-                "duration_ms": duration_ms,
-                "updated_at": now.isoformat(),
-            }
-
-            job.step_status_json = json.dumps(step_status)
-            job.updated_at = now
-            session.add(job)
-            session.commit()
-            return True
-
-    def finalize_postprocess_job(
-        self,
-        job_id: str,
-        outcome: str,
-        error_class: str = None,
-        error_message: str = None,
-    ) -> bool:
-        """Mark a postprocess job as complete.
-
-        Args:
-            job_id: Job identifier
-            outcome: Outcome code (success, failed, skipped_stale)
-            error_class: Exception class name if failed
-            error_message: Error details if failed
-
-        Returns:
-            True if finalized successfully, False if job not found.
-        """
-        now = datetime.now()
-        with self._session() as session:
-            statement = select(PostprocessJobModel).where(
-                PostprocessJobModel.job_id == job_id
-            )
-            job = session.exec(statement).first()
-            if not job:
-                return False
-
-            job.status = (
-                PostProcessStatus.SUCCESS
-                if outcome == PostProcessOutcome.SUCCESS
-                else PostProcessStatus.FAILED
-            )
-            job.outcome = outcome
-            job.finished_at = now
-            job.error_class = error_class
-            job.error_message = error_message
-
-            if job.started_at:
-                job.duration_ms = int((now - job.started_at).total_seconds() * 1000)
-
-            job.updated_at = now
-            session.add(job)
-            session.commit()
-
-            # Update metrics
-            if outcome == PostProcessOutcome.SUCCESS:
-                _postprocess_metrics.job_success += 1
-                _postprocess_metrics.total_duration_ms += job.duration_ms or 0
-            elif outcome == PostProcessOutcome.SKIPPED_STALE:
-                _postprocess_metrics.job_skipped_stale += 1
-            else:
-                _postprocess_metrics.job_failed += 1
-
-            return True
-
-    def get_postprocess_job(self, job_id: str) -> Optional[PostprocessJobModel]:
-        """Retrieve a postprocess job by ID."""
-        with self._session() as session:
-            statement = select(PostprocessJobModel).where(
-                PostprocessJobModel.job_id == job_id
-            )
-            return session.exec(statement).first()
-
-    def list_postprocess_jobs(
-        self, chat_id: str, limit: int = 50
-    ) -> List[PostprocessJobModel]:
-        """List postprocess jobs for a specific chat."""
-        with self._session() as session:
-            statement = (
-                select(PostprocessJobModel)
-                .where(PostprocessJobModel.chat_id == chat_id)
-                .order_by(PostprocessJobModel.created_at.desc())
-                .limit(limit)
-            )
-            return session.exec(statement).all()
-
-    def get_postprocess_metrics(self) -> Dict[str, Any]:
-        """Get current postprocess metrics."""
-        return {
-            "snapshot_committed": _postprocess_metrics.snapshot_committed,
-            "snapshot_failed": _postprocess_metrics.snapshot_failed,
-            "job_started": _postprocess_metrics.job_started,
-            "job_success": _postprocess_metrics.job_success,
-            "job_failed": _postprocess_metrics.job_failed,
-            "job_skipped_stale": _postprocess_metrics.job_skipped_stale,
-            "total_duration_ms": _postprocess_metrics.total_duration_ms,
-        }
-
-    def get_retriable_postprocess_jobs(
-        self, max_age_seconds: int = 3600
-    ) -> List[PostprocessJobModel]:
-        """Get postprocess jobs eligible for retry.
-
-        Criteria:
-        - Status is 'failed' (not 'skipped_stale')
-        - attempt < max_attempts
-        - Not recently attempted (to avoid retry storms)
-
-        Args:
-            max_age_seconds: Only retry jobs older than this (default: 1 hour)
-
-        Returns:
-            List of PostprocessJobModel objects eligible for retry.
-        """
-        cutoff = datetime.now() - timedelta(seconds=max_age_seconds)
-
-        with self._session() as session:
-            statement = (
-                select(PostprocessJobModel)
-                .where(PostprocessJobModel.outcome == PostProcessOutcome.FAILED)
-                .where(PostprocessJobModel.attempt < PostprocessJobModel.max_attempts)
-                .where(PostprocessJobModel.updated_at < cutoff)
-                .order_by(PostprocessJobModel.updated_at.asc())
-                .limit(10)
-            )
-            return session.exec(statement).all()
-
-    def prepare_job_for_retry(self, job_id: str) -> bool:
-        """Prepare a job for retry by incrementing attempt count and resetting status.
-
-        Args:
-            job_id: Job identifier
-
-        Returns:
-            True if prepared successfully, False if job not found or not eligible.
-        """
-        now = datetime.now()
-        with self._session() as session:
-            statement = select(PostprocessJobModel).where(
-                PostprocessJobModel.job_id == job_id
-            )
-            job = session.exec(statement).first()
-            if not job:
-                return False
-
-            # Only reset if currently failed or needs retry
-            if job.outcome != PostProcessOutcome.FAILED:
-                return False
-
-            if job.attempt >= job.max_attempts:
-                return False
-
-            job.attempt += 1
-            job.status = PostProcessStatus.PENDING
-            job.outcome = None
-            job.started_at = None
-            job.finished_at = None
-            job.duration_ms = None
-            job.error_class = None
-            job.error_message = None
-            job.step_status_json = None
-            job.updated_at = now
-
-            session.add(job)
-            session.commit()
-            return True
-
-    # -------------------------------------------------------------------------
-    # Plan Operations
-    # -------------------------------------------------------------------------
-
-    def create_plan(
-        self,
-        chat_id: str,
-        objective: str,
-        tasks: List[Dict[str, Any]] = None,
-    ) -> int:
-        """Create or update the single plan for a chat and return its ID."""
-        now = datetime.now()
-        tasks = tasks or []
-
-        with self._session() as session:
-            # Check for existing plan
-            statement = select(PlanModel).where(PlanModel.chat_id == chat_id).limit(1)
-            existing_plan = session.exec(statement).first()
-
-            if existing_plan:
-                plan_id = existing_plan.id
-                existing_plan.objective = objective
-                existing_plan.updated_at = now
-                session.add(existing_plan)
-
-                # Delete existing tasks
-                task_stmt = select(TaskModel).where(TaskModel.plan_id == plan_id)
-                for task in session.exec(task_stmt).all():
-                    session.delete(task)
-            else:
-                # Create new plan
-                new_plan = PlanModel(
-                    chat_id=chat_id,
-                    objective=objective,
-                    created_at=now,
-                    updated_at=now,
-                )
-                session.add(new_plan)
-                session.commit()
-                session.refresh(new_plan)
-                plan_id = new_plan.id
-
-            # Create tasks
-            for task_data in tasks:
-                task = TaskModel(
-                    plan_id=plan_id,
-                    number=task_data.get("number"),
-                    description=task_data.get("description"),
-                    status=task_data.get("status", "pending"),
-                    note=task_data.get("note"),
-                    capabilities=task_data.get("capabilities"),
-                    created_at=now,
-                    updated_at=now,
-                )
-                session.add(task)
-
-            session.commit()
-            return plan_id
-
-    def get_plan(self, chat_id: str) -> Optional[PlanModel]:
-        """Get the latest plan for a specific chat."""
-        with self._session() as session:
-            statement = (
-                select(PlanModel)
-                .where(PlanModel.chat_id == chat_id)
-                .order_by(PlanModel.created_at.desc())
-                .options(selectinload(PlanModel.tasks))
-                .limit(1)
-            )
-            plan = session.exec(statement).first()
-            if not plan:
-                return None
-
-            # Ensure tasks are sorted (SQLModel might not guarantee order in relationship list)
-            plan.tasks.sort(key=lambda t: t.number)
-            return plan
-
-    def get_plan_by_id(self, plan_id: int) -> Optional[PlanModel]:
-        """Fetch a plan and its tasks by plan ID."""
-        with self._session() as session:
-            statement = (
-                select(PlanModel)
-                .where(PlanModel.id == plan_id)
-                .options(selectinload(PlanModel.tasks))
-            )
-            plan = session.exec(statement).first()
-            if not plan:
-                return None
-
-            plan.tasks.sort(key=lambda t: t.number)
-            return plan
-
-    def list_plans(
-        self,
-        chat_id: str,
-        limit: Optional[int] = None,
-    ) -> List[PlanModel]:
-        """Return all plans for a chat ordered by newest first."""
-        with self._session() as session:
-            statement = (
-                select(PlanModel)
-                .where(PlanModel.chat_id == chat_id)
-                .order_by(PlanModel.created_at.desc())
-                .options(selectinload(PlanModel.tasks))
-            )
-            if limit is not None:
-                statement = statement.limit(limit)
-
-            plans = session.exec(statement).all()
-            for plan in plans:
-                plan.tasks.sort(key=lambda t: t.number)
-            return plans
-
-    def update_plan_objective(self, plan_id: int, objective: str) -> bool:
-        """Update the objective of a plan."""
-        with self._session() as session:
-            plan = session.get(PlanModel, plan_id)
-            if not plan:
-                return False
-
-            plan.objective = objective
-            plan.updated_at = datetime.now()
-            session.add(plan)
-            session.commit()
-            return True
-
-    def create_task(self, plan_id: int, description: str, number: int) -> Optional[int]:
-        """Add a new task to a plan."""
-        now = datetime.now()
-        with self._session() as session:
-            plan = session.get(PlanModel, plan_id)
-            if not plan:
-                return None
-
-            task = TaskModel(
-                plan_id=plan_id,
-                description=description,
-                number=number,
-                created_at=now,
-                updated_at=now,
-            )
-            session.add(task)
-            session.commit()
-            session.refresh(task)
-            return task.id
-
-    def update_task_status(
-        self,
-        chat_id: str,
-        task_number: int,
-        status: str,
-        note: str = None,
-        plan_id: Optional[int] = None,
-    ) -> bool:
-        """Update the status and optionally note of a specific task."""
-        now = datetime.now()
-
-        with self._session() as session:
-            # Find the plan
-            if plan_id is not None:
-                plan = session.get(PlanModel, plan_id)
-                if plan and plan.chat_id != chat_id:
-                    plan = None
-            else:
-                statement = (
-                    select(PlanModel)
-                    .where(PlanModel.chat_id == chat_id)
-                    .order_by(PlanModel.created_at.desc())
-                    .limit(1)
-                )
-                plan = session.exec(statement).first()
-
-            if not plan:
-                return False
-
-            # Find and update the task
-            task_stmt = select(TaskModel).where(
-                (TaskModel.plan_id == plan.id) & (TaskModel.number == task_number)
-            )
-            task = session.exec(task_stmt).first()
-
-            if not task:
-                return False
-
-            task.status = status
-            task.updated_at = now
-            if note is not None:
-                task.note = note
-
-            session.add(task)
-            session.commit()
-            return True
-
-    def update_task(
-        self,
-        task_id: int,
-        status: str = None,
-        description: str = None,
-        note: str = None,
-        capabilities: str = None,
-    ) -> bool:
-        """Update a task's details."""
-        with self._session() as session:
-            task = session.get(TaskModel, task_id)
-            if not task:
-                return False
-
-            if status:
-                task.status = status
-            if description:
-                task.description = description
-            if note:
-                task.note = note
-            if capabilities:
-                task.capabilities = capabilities
-
-            task.updated_at = datetime.now()
-            session.add(task)
-            session.commit()
-            return True
-
-    def delete_task(self, task_id: int) -> bool:
-        """Delete a task."""
-        with self._session() as session:
-            task = session.get(TaskModel, task_id)
-            if not task:
-                return False
-            session.delete(task)
-            session.commit()
-            return True
-
-    def delete_plan(self, chat_id: str) -> bool:
-        """Delete the plan for a specific chat."""
-        with self._session() as session:
-            statement = select(PlanModel).where(PlanModel.chat_id == chat_id)
-            plans = session.exec(statement).all()
-
-            if not plans:
-                return False
-
-            for plan in plans:
-                session.delete(plan)
-
-            session.commit()
-            return True
-
-    # -------------------------------------------------------------------------
-    # User Preferences Operations
-    # -------------------------------------------------------------------------
-
-    def get_user_preferences(self) -> Optional[UserPreferencesModel]:
-        """Get user preferences from the user config file."""
-        from suzent.core.user_config import UserConfigStore
-
-        prefs = UserConfigStore().get_user_preferences()
-        if not prefs:
-            return None
-        return UserPreferencesModel(
-            id=1,
-            model=prefs.get("model"),
-            agent=prefs.get("agent"),
-            tools=prefs.get("tools"),
-            memory_enabled=bool(prefs.get("memory_enabled", False)),
-            sandbox_enabled=bool(prefs.get("sandbox_enabled", True)),
-            sandbox_volumes=prefs.get("sandbox_volumes"),
-            updated_at=self._parse_config_datetime(prefs.get("updated_at")),
-        )
-
-    def save_user_preferences(
-        self,
-        model: str = None,
-        agent: str = None,
-        tools: List[str] = None,
-        memory_enabled: bool = None,
-        sandbox_enabled: bool = None,
-        sandbox_volumes: List[str] = None,
-    ) -> bool:
-        """Save user preferences to the user config file."""
-        from suzent.core.user_config import UserConfigStore
-
-        UserConfigStore().save_user_preferences(
-            {
-                "model": model,
-                "agent": agent,
-                "tools": tools,
-                "memory_enabled": memory_enabled,
-                "sandbox_enabled": sandbox_enabled,
-                "sandbox_volumes": sandbox_volumes,
-            }
-        )
-        return True
-
-    # -------------------------------------------------------------------------
-    # Volume Metadata Operations
-    # -------------------------------------------------------------------------
-
-    def get_volume_metadata(
-        self, volumes: Optional[List[str]] = None
-    ) -> Dict[str, Dict[str, Any]]:
-        """Return cached metadata keyed by volume string."""
-        with self._session() as session:
-            statement = select(VolumeMetadataModel)
-            if volumes is not None:
-                if not volumes:
-                    return {}
-                statement = statement.where(VolumeMetadataModel.volume.in_(volumes))
-
-            rows = session.exec(statement).all()
-            return {
-                row.volume: {
-                    "volume": row.volume,
-                    "host_path": row.host_path,
-                    "mount_point": row.mount_point,
-                    "kind": row.kind,
-                    "exists": row.exists,
-                    "is_git_repo": row.is_git_repo,
-                    "git_root": row.git_root,
-                    "status": row.status,
-                    "error": row.error,
-                    "checked_at": row.checked_at.isoformat()
-                    if row.checked_at
-                    else None,
-                }
-                for row in rows
-            }
-
-    def save_volume_metadata(self, items: List[Dict[str, Any]]) -> bool:
-        """Upsert cached metadata for custom volumes."""
-        now = datetime.now()
-
-        with self._session() as session:
-            for item in items:
-                volume = item.get("volume")
-                if not volume:
-                    continue
-
-                row = session.get(VolumeMetadataModel, volume)
-                if row is None:
-                    row = VolumeMetadataModel(
-                        volume=volume,
-                        host_path=item.get("host_path", ""),
-                        mount_point=item.get("mount_point", ""),
-                        kind=item.get("kind", "generic"),
-                        exists=bool(item.get("exists", False)),
-                        is_git_repo=item.get("is_git_repo"),
-                        git_root=item.get("git_root"),
-                        status=item.get("status", "unknown"),
-                        error=item.get("error"),
-                        checked_at=item.get("checked_at") or now,
-                    )
-                else:
-                    row.host_path = item.get("host_path", "")
-                    row.mount_point = item.get("mount_point", "")
-                    row.kind = item.get("kind", "generic")
-                    row.exists = bool(item.get("exists", False))
-                    row.is_git_repo = item.get("is_git_repo")
-                    row.git_root = item.get("git_root")
-                    row.status = item.get("status", "unknown")
-                    row.error = item.get("error")
-                    row.checked_at = item.get("checked_at") or now
-
-                session.add(row)
-
-            session.commit()
-            return True
-
-    # -------------------------------------------------------------------------
-    # Memory Configuration Operations
-    # -------------------------------------------------------------------------
-
-    def get_memory_config(self) -> Optional[MemoryConfigModel]:
-        """Get memory system configuration from the user config file."""
-        from suzent.core.user_config import UserConfigStore
-
-        config = UserConfigStore().get_memory_config()
-        if not config:
-            return None
-        return MemoryConfigModel(
-            id=1,
-            embedding_model=config.get("embedding_model"),
-            extraction_model=config.get("extraction_model"),
-            updated_at=self._parse_config_datetime(config.get("updated_at")),
-        )
-
-    def save_memory_config(
-        self,
-        embedding_model: str = None,
-        extraction_model: str = None,
-    ) -> bool:
-        """Save memory system configuration to the user config file."""
-        from suzent.core.user_config import UserConfigStore
-
-        UserConfigStore().save_memory_config(
-            {
-                "embedding_model": embedding_model,
-                "extraction_model": extraction_model,
-            }
-        )
-        return True
-
-    # -------------------------------------------------------------------------
-    # MCP Server Operations
-    # -------------------------------------------------------------------------
-
-    def get_mcp_servers(self) -> List[MCPServerModel]:
-        """Get all MCP servers from the database."""
-        with self._session() as session:
-            statement = select(MCPServerModel)
-            servers = session.exec(statement).all()
-            return servers
-
-    def add_mcp_server(
-        self,
-        name: str,
-        config: Dict[str, Any],
-        enabled: bool = True,
-    ) -> bool:
-        """Add a new MCP server configuration."""
-        now = datetime.now()
-        with self._session() as session:
-            if session.get(MCPServerModel, name):
-                return False
-
-            server = MCPServerModel(
-                name=name,
-                type=config.get("type", "stdio"),
-                url=config.get("url"),
-                headers=config.get("headers"),
-                command=config.get("command"),
-                args=config.get("args"),
-                env=config.get("env"),
-                enabled=enabled,
-                created_at=now,
-                updated_at=now,
-            )
-            session.add(server)
-            session.commit()
-            return True
-
-    def update_mcp_server(
-        self, name: str, config: Dict[str, Any] = None, enabled: bool = None
-    ) -> bool:
-        """Update an existing MCP server configuration."""
-        with self._session() as session:
-            server = session.get(MCPServerModel, name)
-            if not server:
-                return False
-
-            if config:
-                if "type" in config:
-                    server.type = config["type"]
-                if "url" in config:
-                    server.url = config["url"]
-                if "headers" in config:
-                    server.headers = config["headers"]
-                if "command" in config:
-                    server.command = config["command"]
-                if "args" in config:
-                    server.args = config["args"]
-                if "env" in config:
-                    server.env = config["env"]
-
-            if enabled is not None:
-                server.enabled = enabled
-
-            server.updated_at = datetime.now()
-            session.add(server)
-            session.commit()
-            return True
-
-    def remove_mcp_server(self, name: str) -> bool:
-        """Remove an MCP server configuration."""
-        with self._session() as session:
-            server = session.get(MCPServerModel, name)
-            if not server:
-                return False
-            session.delete(server)
-            session.commit()
-            return True
-
-    def set_mcp_server_enabled(self, name: str, enabled: bool) -> bool:
-        """Enable or disable an MCP server."""
-        with self._session() as session:
-            server = session.get(MCPServerModel, name)
-            if not server:
-                return False
-
-            server.enabled = enabled
-            server.updated_at = datetime.now()
-            session.add(server)
-            session.commit()
-            return True
-
-    # -------------------------------------------------------------------------
-    # Cron Job Operations
-    # -------------------------------------------------------------------------
-
-    def list_cron_jobs(self, active_only: bool = False) -> List[CronJobModel]:
-        """List all cron jobs, optionally filtered to active only."""
-        with self._session() as session:
-            statement = select(CronJobModel).order_by(CronJobModel.created_at.desc())
-            if active_only:
-                statement = statement.where(CronJobModel.active.is_(True))  # noqa: E712
-            return session.exec(statement).all()
-
-    def get_cron_job(self, job_id: int) -> Optional[CronJobModel]:
-        """Get a cron job by ID."""
-        with self._session() as session:
-            return session.get(CronJobModel, job_id)
-
-    def create_cron_job(
-        self,
-        name: str,
-        cron_expr: str,
-        prompt: str,
-        active: bool = True,
-        delivery_mode: str = "announce",
-        model_override: Optional[str] = None,
-    ) -> int:
-        """Create a new cron job and return its ID."""
-        now = datetime.now()
-        job = CronJobModel(
-            name=name,
-            cron_expr=cron_expr,
-            prompt=prompt,
-            active=active,
-            delivery_mode=delivery_mode,
-            model_override=model_override,
-            created_at=now,
-            updated_at=now,
-        )
-        with self._session() as session:
-            session.add(job)
-            session.commit()
-            session.refresh(job)
-            return job.id
-
-    def update_cron_job(self, job_id: int, **kwargs) -> bool:
-        """Update a cron job's configuration fields."""
-        with self._session() as session:
-            job = session.get(CronJobModel, job_id)
-            if not job:
-                return False
-            for key, value in kwargs.items():
-                if hasattr(job, key) and key not in ("id", "created_at"):
-                    setattr(job, key, value)
-            job.updated_at = datetime.now()
-            session.add(job)
-            session.commit()
-            return True
-
-    def delete_cron_job(self, job_id: int) -> bool:
-        """Delete a cron job by ID."""
-        with self._session() as session:
-            job = session.get(CronJobModel, job_id)
-            if not job:
-                return False
-            session.delete(job)
-            session.commit()
-            return True
-
-    def update_cron_job_run_state(
-        self,
-        job_id: int,
-        last_run_at: Optional[datetime] = None,
-        next_run_at: Optional[datetime] = None,
-        last_result: Optional[str] = None,
-        last_error: Optional[str] = None,
-        retry_count: Optional[int] = None,
-    ) -> bool:
-        """Update cron job run metadata (used by scheduler)."""
-        with self._session() as session:
-            job = session.get(CronJobModel, job_id)
-            if not job:
-                return False
-            if last_run_at is not None:
-                job.last_run_at = last_run_at
-            if next_run_at is not None:
-                job.next_run_at = next_run_at
-            if last_result is not None:
-                job.last_result = last_result[:2000]
-            if last_error is not None:
-                job.last_error = last_error[:1000]
-            if retry_count is not None:
-                job.retry_count = retry_count
-            job.updated_at = datetime.now()
-            session.add(job)
-            session.commit()
-            return True
-
-    # -------------------------------------------------------------------------
-    # Cron Run History
-    # -------------------------------------------------------------------------
-
-    def create_cron_run(self, job_id: int, started_at: datetime) -> int:
-        """Create a run history record, return its ID."""
-        run = CronRunModel(job_id=job_id, started_at=started_at)
-        with self._session() as session:
-            session.add(run)
-            session.commit()
-            session.refresh(run)
-            return run.id
-
-    def finish_cron_run(
-        self,
-        run_id: int,
-        status: str,
-        result: Optional[str] = None,
-        error: Optional[str] = None,
-    ):
-        """Mark a run as finished."""
-        with self._session() as session:
-            run = session.get(CronRunModel, run_id)
-            if not run:
-                return
-            run.finished_at = datetime.now()
-            run.status = status
-            if result is not None:
-                run.result = result[:2000]
-            if error is not None:
-                run.error = error[:1000]
-            session.add(run)
-            session.commit()
-
-    def list_cron_runs(self, job_id: int, limit: int = 20) -> List[CronRunModel]:
-        """List recent runs for a job."""
-        with self._session() as session:
-            statement = (
-                select(CronRunModel)
-                .where(CronRunModel.job_id == job_id)
-                .order_by(CronRunModel.started_at.desc())
-                .limit(limit)
-            )
-            return session.exec(statement).all()
-
-    # -------------------------------------------------------------------------
-    # API Key Operations
-    # -------------------------------------------------------------------------
-
-    def get_api_keys(self) -> Dict[str, str]:
-        """Get non-secret config blobs formerly stored with API keys."""
-        from suzent.core.user_config import UserConfigStore
-
-        return UserConfigStore().get_config_blobs()
-
-    def get_api_key(self, key: str) -> Optional[str]:
-        """Get a config blob or secret by key name."""
-        if key.startswith("_"):
-            from suzent.core.user_config import UserConfigStore
-
-            return UserConfigStore().get_config_blob(key)
-
-        from suzent.core.secrets import get_secret_manager
-
-        return get_secret_manager().get(key)
-
-    def save_api_key(self, key: str, value: str) -> bool:
-        """Save a config blob or secret outside the runtime database."""
-        if key.startswith("_"):
-            from suzent.core.user_config import UserConfigStore
-
-            UserConfigStore().save_config_blob(key, value)
-        else:
-            from suzent.core.secrets import get_secret_manager
-
-            get_secret_manager().set_backend_only(key, value)
-        return True
-
-    def delete_api_key(self, key: str) -> bool:
-        """Delete a config blob or secret outside the runtime database."""
-        if key.startswith("_"):
-            from suzent.core.user_config import UserConfigStore
-
-            UserConfigStore().delete_config_blob(key)
-        else:
-            from suzent.core.secrets import get_secret_manager
-
-            get_secret_manager().delete(key)
-        return True
-
-
-# Global database instance
-_db_instance = None
-
-
-def get_database() -> ChatDatabase:
-    """Get the global database instance."""
-    global _db_instance
-    if _db_instance is None:
-        db_path = os.getenv("CHATS_DB_PATH", None)
-        _db_instance = ChatDatabase(db_path)
-    return _db_instance
-
-
-def generate_chat_title(first_message: str, max_length: int = 50) -> str:
-    """Generate a chat title from the first user message."""
-    if not first_message.strip():
-        return "New Chat"
-
-    title = first_message.strip()
-    title = " ".join(title.split())
-
-    if len(title) > max_length:
-        title = title[: max_length - 3] + "..."
-
-    return title
